Week 2 done

+++ Semaine 2/exercice1.c
@@ -13,12 +13,12 @@ int main() {
 
     disc = b*b-4*a*c;
     if (disc>0) {
-        root1 = (-b - sqrt(disc)) / 2*a; 
-        root2 = (-b + sqrt(disc)) / 2*a;
+        root1 = (float)(-b - sqrt(disc)) / 2*a; 
+        root2 = (float)(-b + sqrt(disc)) / 2*a;
         printf("The solution for the equation %d*x^2 + %d*x + %d = 0 are : \nx = %d \nx = %d",a,b,c,root1,root2);
     }
     else if (disc == 0) {
-        root1 = -b/(2*a);
+        root1 = -b/(float)(2*a);
     printf("The solution for the equation %d*x^2 + %d*x + %d = 0 is : \nx = %d",a,b,c,root1);
     }
     else
